- w.End inclusive (corresponding point still excluded) - interval: int64->float64 - col/row: int64->int

+++ bow/aggregation/windowstart.go
@@ -3,7 +3,7 @@ package aggregation
 import "git.metronlab.com/backend_libraries/go-bow/bow"
 
 func WindowStart(col string) bow.ColumnAggregation {
-	return bow.NewColumnAggregation(col, bow.Int64,
+	return bow.NewColumnAggregation(col, bow.Float64,
 		func(col int, w bow.Window) (interface{}, error) {
 			return w.Start, nil
 		})
